Add simple_rl_helpers
